Re-wrote shouldbet - also with an explanation of why its like that. Also changed howHigh to conform with shouldbet.
Documented both methods. Played about 30-40 rounds and all seems OK. (hamsa hamsa hamsa!)
Check it out, if it seems good to you, delete the old block of shouldbet that you wrote (its in remark now) and tell me what else we need to do.

The way I see it, all we have left is to document the few methods in PlayerStatistics, and write a short document on how the "smart" algorithms work.